+++ tests/misc/fallback_ast_src/php71_or_newer/multi_catch.php
@@ -1,0 +1,6 @@
+<?php
+namespace NS;
+
+try {
+} catch (Exception|\Error|namespace\MyClass $e) {
+}